move NewNBContractConfiguration to nbcontracthelper package
